Move noderepostiory and orchestrator under configserver. No functional changes
